:fire: Rename Java package name
