+++ README.md
@@ -49,7 +49,7 @@ pod install
 #### Android
 
 1. Open up `android/app/src/main/java/[...]/MainApplication.java`
-  - Add `import com.reactlibrary.PDFViewPackage;` to the imports at the top of the file
+  - Add `import com.rumax.reactnative.pdfviewer.PDFViewPackage;` to the imports at the top of the file
   - Add `new PDFViewPackage()` to the list returned by the `getPackages()` method
 2. Append the following lines to `android/settings.gradle`:
     ```
