+++ demo/android/app/src/main/java/com/demo/MainApplication.java
@@ -7,7 +7,7 @@ import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.shell.MainReactPackage;
 import com.facebook.soloader.SoLoader;
-import com.reactlibrary.PDFViewPackage;
+import com.rumax.reactnative.pdfviewer.PDFViewPackage;
 
 import java.util.Arrays;
 import java.util.List;
